backend/requirements.txt
Django==4.2.28
djangorestframework==3.16.1
mysqlclient==2.2.7
python-dotenv==1.2.1
django-cors-headers==4.9.0
google-generativeai==0.3.2
pillow==10.2.0
